+++ IBTool/internal/functions/New-OrgSegment.ps1
@@ -42,13 +42,18 @@
 
 	$statusBar.Text = "Running..."
 	try {
-		Write-PSFHostColor -String "[$((Get-Date).ToString("HH:mm:ss"))] Creating new Organization Segment '$Name'."
-		New-OrganizationSegment -Name $Name -UserGroupFilter "$GroupFilter -$comp '$AttributeValue'" -errorAction Stop
-		Write-PSFHostColor -String "[$((Get-Date).ToString("HH:mm:ss"))] Successfully created Organization Segment '$Name'."
-		$statusBar.Text = "Ready. Created Organization Segment '$Name'."
+		Write-PSFHostColor -String "[$((Get-Date).ToString("HH:mm:ss"))] Creating / editing Organization Segment '$Name'."
+		if ( Get-OrganizationSegment -identity $Name ) {
+			Set-OrganizationSegment -identity $name -UserGroupFilter "$GroupFilter -$comp '$AttributeValue'" -ErrorAction Stop
+		}
+		else {
+			New-OrganizationSegment -Name $Name -UserGroupFilter "$GroupFilter -$comp '$AttributeValue'" -errorAction Stop
+		}
+		Write-PSFHostColor -String "[$((Get-Date).ToString("HH:mm:ss"))] Successfully created / modified Organization Segment '$Name'."
+		$statusBar.Text = "Ready. Created / edited Organization Segment '$Name'."
 	}
 	catch {
-		Write-PSFHostColor -String "[$((Get-Date).ToString("HH:mm:ss"))] Something failed to create the new Organization Segment '$Name'. $_"
-		$statusBar.Text = "Ready. Someting failed to create the new Organization Segment '$Name'. Please see the Powershell window to verify error message."
+		Write-PSFHostColor -String "[$((Get-Date).ToString("HH:mm:ss"))] Something failed to create / edit the Organization Segment '$Name'. $_"
+		$statusBar.Text = "Ready. Someting failed to create / edit the Organization Segment '$Name'. Please see the Powershell window to verify error message."
 	}
 }