+++ IBTool/internal/functions/Start-IBpolicyApp.ps1
@@ -1,5 +1,4 @@
-﻿function Start-IBpolicyApp
-{
+﻿function Start-IBpolicyApp {
 	<#
 	.SYNOPSIS
 	Function to start Information Barrier Policies application to users
@@ -19,6 +18,8 @@
 	)
 	$statusBar.Text = "Running..."
 	try {
+		Write-PSFHostColor -String "[$((Get-Date).ToString("HH:mm:ss"))] Setting all IB Policies status to 'Active'."
+		Get-InformationBarrierPolicy | ForEach-Object { Set-InformationBarrierPolicy -Identity $_.name -State active -Confirm:$false -force -WarningAction SilentlyContinue }
 		Write-PSFHostColor -String "[$((Get-Date).ToString("HH:mm:ss"))] Starting to apply Information Barrier Policies application."
 		Start-InformationBarrierPoliciesApplication -ErrorAction Stop
 		Write-PSFHostColor -String "[$((Get-Date).ToString("HH:mm:ss"))] Started successfully. Allow 30 minutes for the system to start applying the policies. The system applies policies user by user. The system processes about 5,000 user accounts per hour."